Fix a bug in sort.cc and gklib.cc
